all daos are working except CadministratorNotesDao tests, created a separate package for scripts daos
