+++ src/main/java/org/mehaexample/asdDemo/script/TopFiveEmployersDao.java
@@ -1,4 +1,4 @@
-package org.mehaexample.asdDemo.dao;
+package org.mehaexample.asdDemo.script;
 
 import org.mehaexample.asdDemo.Constants;
