+++ src/main/java/org/mehaexample/asdDemo/script/TotalStudentsInGoogleDao.java
@@ -1,5 +1,4 @@
-package org.mehaexample.asdDemo.dao;
-
+package org.mehaexample.asdDemo.script;
 
 import java.sql.SQLException;
